+++ src/Player/pause/player-pause.tsx
