+++ src/index.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom/client';
-import {App} from './App.tsx';
+import {App} from './app.tsx';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
