+++ src/movie-page/in-list/movie-page-in-list.tsx
